+++ CHANGELOG.md
@@ -21,6 +21,8 @@ All notable changes to this project will be documented in this file.
     - Jungle to Lost World.
     - Jungle to Lost World (Alt).
     - Jungle to Sand Hill.
+- You are now allowed to pause while the game is displaying subtitles.
+  - This won't affect the traps that much, but it can help with key items. 
 
 ## [v1.2.1] - 2026-??-??